Fix -Wimplicit-float-conversion
Fixes include:
- double <-> float in arithmetic, descriptor passing, comparison.
- int <-> float in arithmetic, descriptor passing, comparison,
  string formatting.

that are mostly harmless.

(ResidencyManagerD3D12.cpp change promotes to double for big int
math that should increase precision)

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/326575
Commit-Queue: Shrek Shao <shrekshao@google.com>

+++ src/dawn/native/CopyTextureForBrowserHelper.cpp
@@ -399,8 +399,9 @@ MaybeError DoCopyForBrowser(DeviceBase* device,
     // this flip when converting positions to texcoords.
     // https://www.w3.org/TR/webgpu/#coordinate-systems
     if (!options->flipY) {
-        uniformData.scaleY *= -1.0;
-        uniformData.offsetY += copySize->height / static_cast<float>(sourceInfo->size.height);
+        uniformData.scaleY *= -1.0f;
+        uniformData.offsetY +=
+            static_cast<float>(copySize->height) / static_cast<float>(sourceInfo->size.height);
     }
 
     uint32_t stepsMask = 0u;
