+++ src/tint/lang/wgsl/resolver/eval_test.h
@@ -145,7 +145,9 @@ inline void CheckConstant(const constant::Value* got_constant,
                                             static_cast<double>(*flags.float_compare_epsilon))
                                     << "index: " << i;
                             } else {
-                                EXPECT_FLOAT_EQ(got, expected) << "index: " << i;
+                                EXPECT_FLOAT_EQ(static_cast<float>(got),
+                                                static_cast<float>(expected))
+                                    << "index: " << i;
                             }
                         } else {
                             EXPECT_EQ(got, expected) << "index: " << i;
